Update dependencies to 1.18.1 and fix compilation-level problems
