+++ common/src/main/java/hardcorequesting/common/tileentity/AbstractBarrelBlockEntity.java
@@ -10,6 +10,7 @@ import hardcorequesting.common.quests.task.item.ConsumeItemTask;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
+import net.minecraft.network.protocol.game.ClientboundBlockEntityDataPacket;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -116,6 +117,16 @@ public abstract class AbstractBarrelBlockEntity extends AbstractBaseBlockEntity 
                 }
             }
         }
+    }
+    
+    @Override
+    public ClientboundBlockEntityDataPacket getUpdatePacket() {
+        return ClientboundBlockEntityDataPacket.create(this);
+    }
+    
+    @Override
+    public CompoundTag getUpdateTag() {
+        return saveWithoutMetadata();
     }
     
     protected void updateState() {
