+++ fabric/gradle.properties
@@ -1,8 +1,8 @@
 
-fabric_version=0.40.8+1.17
+fabric_version=0.44.0+1.18
 
-architectury_version=2.5.28
+architectury_version=3.2.51
 
-roughlyenoughitems=6.0.272-alpha
-cardinal_components=3.1.1
-lba=0.9.0
+roughlyenoughitems=7.0.352
+cardinal_components=4.0.1
+lba=0.10.0
