+++ forge/gradle.properties
@@ -1,3 +1,3 @@
 loom.platform=forge
 
-forge_version=37.0.44
+forge_version=39.0.5
