+++ gradle.properties
@@ -4,6 +4,6 @@ mod_version=5.8.0
 maven_group=me.shedaniel
 archives_base_name=HQM
 
-minecraft_version=1.17.1
+minecraft_version=1.18.1
 
-fabric_loader_version=0.11.6
+fabric_loader_version=0.12.12
